address recent comment on PrevFilePaths utility and others
