+++ hudi-client/hudi-flink-client/src/main/java/org/apache/hudi/client/HoodieFlinkTableServiceClient.java
@@ -100,7 +100,6 @@ public class HoodieFlinkTableServiceClient<T> extends BaseHoodieTableServiceClie
         }
       }
       log.info("Compacted successfully on commit {}", compactionCommitTime);
-      log.info("Compacted successfully on commit " + compactionCommitTime);
       fireCommitCallbackIfNecessary(compactionCommitTime, HoodieActiveTimeline.COMMIT_ACTION,
           metadata.getWriteStats(), table::getBaseFileOnlyView, Option.empty());
     } finally {
@@ -163,7 +162,6 @@ public class HoodieFlinkTableServiceClient<T> extends BaseHoodieTableServiceClie
       }
     }
     log.info("Clustering successfully on commit {}", clusteringCommitTime);
-    log.info("Clustering successfully on commit " + clusteringCommitTime);
     fireCommitCallbackIfNecessary(clusteringCommitTime, clusteringInstant.getAction(),
         writeStats, table::getBaseFileOnlyView, Option.empty());
   }
